Fixes hang when stopping MATLAB from matlab-proxy via interrupt signal.
This issue usually only occurs when MATLAB starts MathWorksServiceHost process for the first time through matlab-proxy. In this state, the matlab-proxy process shares a pipe with the MathWorksServiceHost process and waits for ever for that pipe to close. This change closes the shared stderr pipe before waiting for MATLAB to terminate.